fairgen/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.8.0
google-genai>=1.0.0
google-cloud-aiplatform>=1.60.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.16.0
scikit-learn>=1.5.0
datasets>=2.20.0
huggingface_hub>=0.24.0
sdv>=1.17.0
python-dotenv>=1.0.1
gspread>=6.0.0
google-auth>=2.28.0
